wallet controller fix
